fix(suite): implement remaining trezor models and separate chunk logic
